# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VetControl API — veterinary clinic management system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pacientes:
    get:
      operationId: getPacientes
      tags: [pacientes]
      summary: List patients with optional search
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PacienteListItem"
    post:
      operationId: createPaciente
      tags: [pacientes]
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PacienteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paciente"
  /pacientes/{pacienteId}:
    get:
      operationId: getPaciente
      tags: [pacientes]
      summary: Get patient detail with full clinical history
      parameters:
        - name: pacienteId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Patient detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PacienteDetalle"
        "404":
          description: Not found
    put:
      operationId: updatePaciente
      tags: [pacientes]
      summary: Update a patient
      parameters:
        - name: pacienteId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PacienteInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paciente"
        "404":
          description: Not found
    delete:
      operationId: deletePaciente
      tags: [pacientes]
      summary: Delete a patient
      parameters:
        - name: pacienteId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /pacientes/{pacienteId}/recetas:
    post:
      operationId: createReceta
      tags: [recetas]
      summary: Create a prescription for a patient
      parameters:
        - name: pacienteId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecetaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecetaDetalle"
        "404":
          description: Not found
  /recetas/{recetaId}:
    get:
      operationId: getReceta
      tags: [recetas]
      summary: Get a prescription with its items
      parameters:
        - name: recetaId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Prescription detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecetaDetalle"
        "404":
          description: Not found
    delete:
      operationId: deleteReceta
      tags: [recetas]
      summary: Delete a prescription
      parameters:
        - name: recetaId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /pacientes/{pacienteId}/pruebas:
    post:
      operationId: createPrueba
      tags: [pruebas]
      summary: Create a diagnostic test for a patient
      parameters:
        - name: pacienteId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PruebaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prueba"
        "404":
          description: Not found
  /pruebas/{pruebaId}:
    get:
      operationId: getPrueba
      tags: [pruebas]
      summary: Get a diagnostic test
      parameters:
        - name: pruebaId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Diagnostic test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prueba"
        "404":
          description: Not found
    put:
      operationId: updatePrueba
      tags: [pruebas]
      summary: Update a diagnostic test
      parameters:
        - name: pruebaId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PruebaInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prueba"
        "404":
          description: Not found
    delete:
      operationId: deletePrueba
      tags: [pruebas]
      summary: Delete a diagnostic test
      parameters:
        - name: pruebaId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
...
    Paciente:
      type: object
      properties:
        id:
          type: number
        clienteId:
          type: number
        nombre:
          type: string
        especie:
          type: string
        raza:
          type: ["string", "null"]
        sexo:
          type: ["string", "null"]
        fechaNacimiento:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        peso:
          type: ["number", "null"]
        microchip:
          type: ["string", "null"]
        esterilizado:
          type: boolean
        estado:
          type: string
        alergias:
          type: ["string", "null"]
        antecedentes:
          type: ["string", "null"]
        vacunas:
          type: ["string", "null"]
        alimentacion:
          type: ["string", "null"]
        creadoEn:
          type: string
      required:
        - id
        - clienteId
        - nombre
        - especie
        - esterilizado
        - estado
        - creadoEn
    PacienteListItem:
      type: object
      properties:
        id:
          type: number
        nombre:
          type: string
        especie:
          type: string
        raza:
          type: ["string", "null"]
        propietario:
          type: string
        clienteId:
          type: number
        peso:
          type: ["number", "null"]
        estado:
          type: string
        creadoEn:
          type: string
      required:
        - id
        - nombre
        - especie
        - propietario
        - clienteId
        - estado
        - creadoEn
    PacienteDetalle:
      type: object
      properties:
        paciente:
          $ref: "#/components/schemas/Paciente"
        propietario:
          type: string
        propietarioId:
          type: number
        consultas:
          type: array
          items:
            $ref: "#/components/schemas/Consulta"
        recetas:
          type: array
          items:
            $ref: "#/components/schemas/RecetaDetalle"
        pruebas:
          type: array
          items:
            $ref: "#/components/schemas/Prueba"
      required:
        - paciente
        - propietario
        - propietarioId
        - consultas
        - recetas
        - pruebas
    PacienteInput:
      type: object
      properties:
        clienteId:
          type: number
        nombre:
          type: string
          minLength: 1
        especie:
          type: string
          minLength: 1
        raza:
          type: string
        sexo:
          type: string
        fechaNacimiento:
          type: string
        color:
          type: string
        peso:
          type: number
        microchip:
          type: string
        esterilizado:
          type: boolean
        estado:
          type: string
        alergias:
          type: string
        antecedentes:
          type: string
        vacunas:
          type: string
        alimentacion:
          type: string
      required:
        - clienteId
        - nombre
        - especie
    Consulta:
      type: object
      properties:
        id:
          type: number
        pacienteId:
          type: number
        fecha:
          type: string
        hora:
          type: ["string", "null"]
        medico:
          type: ["string", "null"]
        motivo:
          type: string
        anamnesis:
          type: ["string", "null"]
        exploracionFisica:
          type: ["string", "null"]
        peso:
          type: ["number", "null"]
        temperatura:
          type: ["number", "null"]
        frecuenciaCardiaca:
          type: ["number", "null"]
        frecuenciaRespiratoria:
          type: ["number", "null"]
        mucosas:
          type: ["string", "null"]
        trc:
          type: ["string", "null"]
        condicionCorporal:
          type: ["string", "null"]
        diagnostico:
          type: ["string", "null"]
        diagnosticosDiferenciales:
          type: ["string", "null"]
        plan:
          type: ["string", "null"]
        tratamiento:
          type: ["string", "null"]
        pronostico:
          type: ["string", "null"]
        proximaCita:
          type: ["string", "null"]
        observaciones:
          type: ["string", "null"]
        creadoEn:
          type: string
      required:
        - id
        - pacienteId
        - fecha
        - motivo
        - creadoEn
...
    RecetaPartida:
      type: object
      properties:
        id:
          type: number
        recetaId:
          type: number
        medicamento:
          type: string
        presentacion:
          type: ["string", "null"]
        dosis:
          type: string
        via:
          type: ["string", "null"]
        frecuencia:
          type: ["string", "null"]
        duracion:
          type: ["string", "null"]
        instrucciones:
          type: ["string", "null"]
      required:
        - id
        - recetaId
        - medicamento
        - dosis
    RecetaPartidaInput:
      type: object
      properties:
        medicamento:
          type: string
          minLength: 1
        presentacion:
          type: string
        dosis:
          type: string
          minLength: 1
        via:
          type: string
        frecuencia:
          type: string
        duracion:
          type: string
        instrucciones:
          type: string
      required:
        - medicamento
        - dosis
    RecetaDetalle:
      type: object
      properties:
        id:
          type: number
        pacienteId:
          type: number
        consultaId:
          type: ["number", "null"]
        fecha:
          type: string
        medico:
          type: ["string", "null"]
        indicacionesGenerales:
          type: ["string", "null"]
        proximaRevision:
          type: ["string", "null"]
        partidas:
          type: array
          items:
            $ref: "#/components/schemas/RecetaPartida"
        paciente:
          type: string
        propietario:
          type: string
        especie:
          type: ["string", "null"]
        raza:
          type: ["string", "null"]
        creadoEn:
          type: string
      required:
        - id
        - pacienteId
        - fecha
        - partidas
        - paciente
        - propietario
        - creadoEn
    RecetaInput:
      type: object
      properties:
        consultaId:
          type: number
        fecha:
          type: string
          minLength: 1
        medico:
          type: string
        indicacionesGenerales:
          type: string
        proximaRevision:
          type: string
        partidas:
          type: array
          items:
            $ref: "#/components/schemas/RecetaPartidaInput"
          minItems: 1
      required:
        - fecha
        - partidas
    Prueba:
      type: object
      properties:
        id:
          type: number
        pacienteId:
          type: number
        consultaId:
          type: ["number", "null"]
        fecha:
          type: string
        tipo:
          type: ["string", "null"]
        nombre:
          type: string
        laboratorio:
          type: ["string", "null"]
        resultado:
          type: ["string", "null"]
        interpretacion:
          type: ["string", "null"]
        costo:
          type: number
        creadoEn:
          type: string
      required:
        - id
        - pacienteId
        - fecha
        - nombre
        - costo
        - creadoEn
    PruebaInput:
      type: object
      properties:
        consultaId:
          type: number
        fecha:
          type: string
          minLength: 1
        tipo:
          type: string
        nombre:
          type: string
          minLength: 1
        laboratorio:
          type: string
        resultado:
          type: string
        interpretacion:
          type: string
        costo:
          type: number
      required:
        - fecha
        - nombre
...