# lib/api-zod/src/generated/types/pacienteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VetControl API — veterinary clinic management system
 * OpenAPI spec version: 0.1.0
 */

export interface PacienteInput {
  clienteId: number;
  /** @minLength 1 */
  nombre: string;
  /** @minLength 1 */
  especie: string;
  raza?: string;
  sexo?: string;
  fechaNacimiento?: string;
  color?: string;
  peso?: number;
  microchip?: string;
  esterilizado?: boolean;
  estado?: string;
  alergias?: string;
  antecedentes?: string;
  vacunas?: string;
  alimentacion?: string;
}