# lib/api-zod/src/generated/types/pruebaInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VetControl API — veterinary clinic management system
 * OpenAPI spec version: 0.1.0
 */

export interface PruebaInput {
  consultaId?: number;
  /** @minLength 1 */
  fecha: string;
  tipo?: string;
  /** @minLength 1 */
  nombre: string;
  laboratorio?: string;
  resultado?: string;
  interpretacion?: string;
  costo?: number;
}