# lib/api-zod/src/generated/types/recetaInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VetControl API — veterinary clinic management system
 * OpenAPI spec version: 0.1.0
 */
import type { RecetaPartidaInput } from './recetaPartidaInput';

export interface RecetaInput {
  consultaId?: number;
  /** @minLength 1 */
  fecha: string;
  medico?: string;
  indicacionesGenerales?: string;
  proximaRevision?: string;
  /** @minItems 1 */
  partidas: RecetaPartidaInput[];
}